生徒側のチャプターの進捗率を取得するprogressメソッドをCourseControllerからAttendanceController内へ移植。 ルーティングを修正し、パラメータとしてattendance_idを渡すだけでprogressメソッドで進捗率が取得できるようにロジックも修正。 progressメソッド内でログインしている生徒と受講データの生徒のidが一致しているときだけ進捗率が取得できるように実装。 AttendanceController内のprogressメソッド用のバリデーションのために新たにRequestクラスを作成しprogressメソッドに反映させた。 AttendanceController内のprogressメソッド用の値を返すために新たにResourceクラスを作成し、progressメソッドでインスタンス化できるように実装。 AttendanceController内で呼び出せるようにCourseController内の5つのメソッドも一緒にAttendanceController内へ移植。 移植完了後に不要になったCourseController内のprogressメソッドと他5つのメソッドを削除。 作り直したことで不要になった元々のRequestクラスとResourceクラスを削除。
