+++ app/Http/Requests/Student/AttendanceCourseProgressRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Student;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class CourseProgressRequest extends FormRequest
+class AttendanceCourseProgressRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -19,7 +19,7 @@ class CourseProgressRequest extends FormRequest
     protected function prepareForValidation()
     {
         $this->merge([
-            'course_id' => $this->route('course_id')
+            'attendance_id' => $this->route('attendance_id')
         ]);
     }
 
@@ -30,8 +30,8 @@ class CourseProgressRequest extends FormRequest
      */
     public function rules()
     {
-        return [
-            'course_id' => ['required', 'integer', 'exists:courses,id']
+        return [    
+            'attendance_id' => ['required', 'integer', 'exists:attendances,id']
         ];
     }
 }
