+++ app/Http/Resources/Student/AttendanceCourseProgressResource.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Resources;
+namespace App\Http\Resources\Student;
 
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class CourseProgressResource extends JsonResource
+class AttendanceCourseProgressResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
